requirements.txt
streamlit
opencv-python
yt-dlp
img2pdf
numpy
fpdf2
Pillow